requirements.txt
apispec==6.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pipenv==2022.10.12
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
webargs==8.2.0
Werkzeug==2.2.2

python-dotenv~=0.21.0